requirements.txt
fastapi~=0.112.2
numpy~=1.26.4
scikit-learn~=1.2.2
pydantic~=1.10.12
uvicorn~=0.20.0
pandas~=2.1.4
pyyaml~=6.0.1